+++ taskjuggler/ReportElement.h
@@ -53,7 +53,7 @@ public:
 
     void addScenario(int sc) { scenarios.append(sc); }
     void clearScenarios() { scenarios.clear(); }
-    uint getScenarioCount() const { return scenarios.count(); }
+    size_t getScenarioCount() const { return scenarios.count(); }
     int getScenario(int sc) const { return scenarios[sc]; }
 
     void setHeadline(const QString& hl) { headline = hl; }
